notoriety tagging

+++ harvest-chart-server/src/schema_types.rs
@@ -24,7 +24,7 @@ pub struct BasePlant {
     pub name_fts: String,
     pub type_: String,
     pub number_of_references: i32,
-    pub notoriety_score: Option<f64>,
+    pub notoriety_score: Option<f32>,
     pub notoriety_score_explanation: Option<String>,
     pub aka: Option<String>,
     pub aka_fts: Option<String>,
@@ -79,7 +79,7 @@ pub struct Collection {
     pub filename: Option<String>,
 
     pub notoriety_type: String,
-    pub notoriety_score: Option<f64>,
+    pub notoriety_score: Option<f32>,
     pub notoriety_score_explanation: Option<String>,
 
     pub title: Option<String>,
